requirements.txt
flask
opencv-python
numpy
tensorflow-cpu==2.9.0
keras
pillow
scikit-learn
werkzeug
gunicorn